Chunk update

+++ include/renderer/renderer.h
@@ -47,9 +47,10 @@ namespace SymoCraft{
         // void draw3DModel(const glm::vec3& position, const glm::vec3& scale, float rotation, const VoxelVertex* vertices, int verticesLength);
         // void drawLine(const glm::vec3& start, const glm::vec3& end, const Style& style);
         // void drawBox(const glm::vec3& block_center_coord, const glm::vec3& size, const Style& style);
-        void AddBlocksToBatch(const glm::vec3 &block_center_coord, const uint16 &side_tex, const uint16 &top_tex, const uint16 &bottom_tex);
+        void AddBlocksToBatch(const glm::ivec3 &block_center_coord, const uint16 &side_tex, const uint16 &top_tex, const uint16 &bottom_tex);
 
         void ReportStatus();
+
     }
 
 }
